# backend/api/docs/swagger.yaml
basePath: /
definitions:
...
  models.Message:
    properties:
      _id:
        type: string
      content:
        type: string
      recever:
        type: string
      sender:
        type: string
    type: object
...
  models.SendMessageM:
    properties:
      content:
        minLength: 5
        type: string
      recever:
        type: string
      sender:
        type: string
    required:
    - content
    - recever
    - sender
    type: object
...
host: localhost:5000
info:
  contact: {}
  description: 这是基于 Golang Fiber 的 REST API Swagger 文档
  title: Fiber Golang Mongo Grpc Websocket 等服务
  version: "1.0"
paths:
  /chat/getmsgsbynums:
    get:
      consumes:
      - application/json
      description: 按页获取两个用户之间的聊天记录
      parameters:
      - description: 起始页码
        in: query
        name: from
        required: true
        type: integer
      - description: 第一个用户 ID
        in: query
        name: firstuid
        required: true
        type: string
      - description: 第二个用户 ID
        in: query
        name: seconduid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Message'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 分页获取聊天消息
      tags:
      - Chat
  /chat/sendmessage:
    post:
      consumes:
      - application/json
      description: 在两个用户之间发送消息
      parameters:
      - description: 消息发送参数
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/models.SendMessageM'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Message'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 发送私信
      tags:
      - Chat
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: 使用 Bearer 认证，格式为 "Bearer 空格 Token"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"